Refactoring: Rename and move classes
